backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
Pillow==10.1.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
numpy==1.26.2
playwright==1.57.0
